+++ src/HIDra.Models/VirtualKey.cs
@@ -95,5 +95,6 @@ public enum VirtualKey : ushort
     BrowserBack = 0xA6,
     VolumeMute = 0xAD,
     VolumeDown = 0xAE,
-    VolumeUp = 0xAF
+    VolumeUp = 0xAF,
+    MediaPlayPause = 0xB3
 }